% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{multiply2}
\alias{multiply2}
\title{An example multiplication function}
\usage{
multiply2(x1, x2)
}
\arguments{
\item{x1}{is the first number in the multiplication}

\item{x2}{is the second number in the multiplication}
}
\description{
An example multiplication function
}
