% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_all_sections.r
\name{scrape_all_sections}
\alias{scrape_all_sections}
\title{Scrape all CDC Vital Statistics sections}
\usage{
scrape_all_sections(url)
}
\arguments{
\item{url}{Character string with the CDC Vital Stats page URL.}
}
\value{
A tibble with all file links and metadata for all sections.
}
\description{
Downloads and combines all the main CDC Vital Statistics sections
into a single tibble. Uses \code{get_html_page()} and \code{scrape_cdc_section()}.
}
\examples{
\dontrun{
all_cdc_data <- scrape_all_sections("https://www.cdc.gov/nchs/data_access/VitalStatsOnline.htm")
}

}
